requirements.txt
# Important Requirements
https://github.com/New-dev0/Telethon/archive/diamond.zip
python-decouple
python-dotenv

# Requirements for Plugins.
akinator
apscheduler
aiohttp
bs4
enhancer>=0.3.4
git+https://github.com/New-dev0/instagrapi.git@39df1b1#egg=instagrapi
git+https://github.com/buddhhu/img2html.git@c44170d#egg=img2html
heroku3
gitpython
google-api-python-client
htmlwebshot
lottie
numpy>=1.21.2
oauth2client
opencv-python-headless
pillow>=7.0.0
profanitydetector
psutil
pypdf2>=1.26.0
pytz
qrcode
requests
tabulate
git+https://github.com/xditya/telegraph.git@41fd800#egg=telegraph
tgcrypto
youtube-search-python
yt-dlp

# Requirements for Addons.
covid
emoji
fonttools
gtts
gingerit
jikanpy
git+https://github.com/New-dev0/AsyncLyricsExtractor.git
phlogo
pokedex.py
pyfiglet
pygments
pyjokes
quotefancy
shazamio
speechrecognition
speedtest-cli
textblob
wikipedia
